test: extract MockScreenCapturePermissionAccess.swift from UtilityTestSupport.swift
Splits `MockScreenCapturePermissionAccess` out of UtilityTestSupport.swift
(the mock covers ScreenCapturePermissionAccessing, not a general
"utility" seam). UtilityTestSupport.swift retains MockHotkeyManager
and MockURLHandler.

Byte-identical move of lines 33-53. Both files retain their test imports.

UtilityTestSupport.swift: 53 → 31 lines. Tests: 667.
